+++ src/main/java/com/susanghan_guys/server/global/security/handler/JwtAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package com.susanghan_guys.server.global.jwt;
+package com.susanghan_guys.server.global.security.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.susanghan_guys.server.global.common.CommonResponse;
@@ -6,29 +6,28 @@ import com.susanghan_guys.server.global.common.code.ErrorCode;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 
 @Component
 @RequiredArgsConstructor
-public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
+public class JwtAccessDeniedHandler implements AccessDeniedHandler {
 
     private final ObjectMapper objectMapper;
 
     @Override
-    public void commence(
+    public void handle(
             HttpServletRequest request,
             HttpServletResponse response,
-            AuthenticationException authException
+            AccessDeniedException accessDeniedException
     ) throws IOException {
-        response.setStatus(HttpStatus.UNAUTHORIZED.value());
+        response.setStatus(HttpServletResponse.SC_FORBIDDEN);
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
         response.setCharacterEncoding("UTF-8");
-        objectMapper.writeValue(response.getWriter(), CommonResponse.failure(ErrorCode.UNAUTHORIZED));
+        objectMapper.writeValue(response.getWriter(), CommonResponse.failure(ErrorCode.FORBIDDEN));
     }
 }
